Cast weapon offset height to int
Fixes hairline cracks between bottom of weapon and HUD at some resolutions (similar to cracked parchment before same fix made).

+++ Assets/Scripts/Game/FPSWeapon.cs
@@ -301,7 +301,7 @@ namespace DaggerfallWorkshop.Game
                 WeaponOffsetHeight = 0;
                 if (DaggerfallUnity.Settings.LargeHUD && DaggerfallUnity.Settings.LargeHUDOffsetWeapon)
                 {
-                    WeaponOffsetHeight = DaggerfallUI.Instance.DaggerfallHUD.LargeHUD.Rectangle.height;
+                    WeaponOffsetHeight = (int)DaggerfallUI.Instance.DaggerfallHUD.LargeHUD.Rectangle.height;
                 }
 
                 // Source weapon images are designed to overlay a fixed 320x200 display.
